+++ packages/vite-plugin-jsx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/vite-plugin-jsx
 
+## 1.2.0
+
+### Minor Changes
+
+- Distinguish between in-scope and out-of-scope primitives in Vite plugins
+
 ## 1.1.1
 
 ### Patch Changes